# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wholesome Uganda API Docs",
    "version": "1.0.0",
    "description": "API documentation for Wholesome Uganda website"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev server"
    },
    {
      "url": "https://api.wholesomeug.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/events/all": {
      "get": {
        "summary": "Get all events (including past events)",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Returns all events including past ones. Admin access required.",
        "responses": {
          "200": {
            "description": "List of all events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/gallery": {
      "get": {
        "summary": "Get all gallery items",
        "tags": [
          "Gallery"
        ],
        "responses": {
          "200": {
            "description": "List of gallery items"
          }
        }
      },
      "post": {
        "summary": "Upload multiple gallery items",
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "Optional description to apply to all uploaded images"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of image files"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Images uploaded successfully"
          },
          "400": {
            "description": "Bad request (e.g. no files or file too large)"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/gallery/{id}": {
      "delete": {
        "summary": "Delete a gallery item",
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Gallery item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/newsletter/subscribe": {
      "post": {
        "summary": "Subscribe to the newsletter",
        "tags": [
          "Newsletter"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscribed successfully"
          },
          "400": {
            "description": "Already subscribed or bad request"
          },
          "429": {
            "description": "Cooldown period not met"
          }
        }
      }
    },
    "/newsletter/unsubscribe": {
      "post": {
        "summary": "Unsubscribe from the newsletter",
        "tags": [
          "Newsletter"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsubscribed successfully"
          },
          "400": {
            "description": "Email not found or already unsubscribed"
          }
        }
      }
    },
    "/newsletter": {
      "get": {
        "summary": "Get all subscribers",
        "tags": [
          "Newsletter"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of subscribed emails"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}